requirements.txt
pandas>=2.2
numpy>=2.0
matplotlib>=3.9
seaborn>=0.13
lifelines>=0.30
scikit-learn>=1.5
feature-engine>=1.8
imbalanced-learn>=0.13
xgboost>=3.2
shap>=0.51
flask>=3.1
streamlit>=1.56
evidently>=0.7
python-dotenv>=1.2
